made code changes: removed getChannel members service function from channelMembers.service.ts to channels.service.ts
